Optimization: Microsoft account login changed to local processing
優化：微軟帳號登入改為本地端處理

+++ lib/Model/Game/ModInfo.dart
@@ -81,7 +81,9 @@ class ModInfo {
                 child: I18nText("gui.confirm"),
                 onPressed: () {
                   Navigator.of(context).pop();
-                  file.deleteSync(recursive: true);
+                  if (file.existsSync()) {
+                    file.deleteSync(recursive: true);
+                  }
                 })
           ],
         );
